Sprint 1 work

+++ src/com/company/DeleteFile.java
@@ -11,25 +11,40 @@ public class DeleteFile {
     //public static String main(String[] args) throws IOException {
         File FileDir = new File("D:\\LockedMe");
         List<String> listFile = Arrays.asList(FileDir.list());
+        File[] filenamelist = FileDir.listFiles();
         if (FileDir.isDirectory()) {
             Scanner FileName = new Scanner(System.in);
+
             System.out.println("Enter the file name wants to Delete With Extension( Like- .txt,.doc etc) :- ");
             String Nufil = FileName.nextLine();
             File Newfile = new File("D:\\LockedMe\\" + Nufil);
+            //System.out.println(Newfile);
+            //boolean test;
+            for (File i : filenamelist) {
+//              System.out.println( i );
+                String FN = i.getName();
+//              String M= FN.toLowerCase();
+//              //String N = Nufil.toLowerCase();
+//               System.out.println(FN);
+//               //test = M.equals(N);
+                if (FN.equals(Nufil)) {
+                    if (Newfile.delete()) {
 
-                       // for (Object S : listFile) {
-            // boolean S = listFile.contains(Nufil);
+                        System.out.println(" Your Have Successfully delete the file - " + Nufil);
+                        return null;
 
-//           if(FileDir.exists() && Newfile.getCanonicalPath().equals(listFile)){
-            if(Newfile.delete()){
-                Collections.sort(listFile) ;
-                System.out.println(" Your Have Successfully delete the file - " + Nufil);
-            } else {
-                System.out.println("File Not Found Please write correct file name with extension ");
-                return DeleteFile();
-            }
+                    }}  }
+            System.out.println("File Not Found Please write correct file name with extension ");
+            return DeleteFile();
+
+//                } else {
+//                    System.out.println("File Not Found Please write correct file name with extension ");
+//                    return DeleteFile();
+//                }
 
         }
+
+
 
         else {
             System.out.println(FileDir.getAbsolutePath() + " Please Give correct Path of Directory or folder ");
